requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.2
django-calculation==1.0.0
django-countries==7.5.1
pillow==10.2.0
reportlab==4.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
